Add PriorityTaskQueue
